Remove interrupt guards in read/write
I initially thought this was a good change, but when I gave it
deeper thought I decided that this change raises two issues:

1. It violates one of my core design philosophies which is "zero
assumptions." On this basis I feel that it is more our
responsibility to simply document and allow the user to disable
interrupts when interacting with the library if they feel it is
appropriate.

2. If using an RTOS or similar framework, disabling all interrupts
from the pico sdk could be a problem. By allowing the user to
disable interrupts when they deem appropriate, they are able to
choose the correct method based on their use case. This really
falls back to "zero assumptions" and I just feel that the
management of any critical systems like interrupts should not be
touched withing low level driver functions unless strictly
necessary.