Add FIFO read/write/clear helpers + cleanup
I felt silly never having added FIFO read/write with how
essential a function it is. I also spotted in the datasheet that
you can write a 1 into the FifoOverrun flag at any time to clear
the FIFO, thus the new fifo_clear function.

I corrected the default sync word length (3 -> 4). The default
register value is 3, but this represents 4 bytes of the default sync
value being used.

I also removed the setting of the preamble length during init as this
is already the default register value on reset.